Add project configurable Room Data Size
This commit also removes MetaTiles::EngineSettings as it is no longer
needed.

+++ src/models/project/rom-data-writer.hpp
@@ -17,7 +17,7 @@ class RomDataWriter {
 public:
     struct Constant {
         const std::string name;
-        const int value;
+        const long value;
     };
 
 private:
